multiple warning fixes with gcc-apple
git-svn-id: file:///home/michael/svn/yuni/trunk@2440 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/messaging/transport/rest/request.inc.hpp
@@ -360,7 +360,7 @@ namespace REST
 			context.method = mhandler.name;
 			context.schema = mhandler.schema;
 			context.httpStatus = 200;
-			context.remotePort = reqinfo.remote_port;
+			context.remotePort = (uint)reqinfo.remote_port;
 			// response
 			Marshal::Object response;
